FLUID-5789: Addressing an issue with clean
The package.json file couldn't be cleaned in windows systems due to issues
with how node and git handle absolute paths differently. This has been addressed
by changing the working directly briefly to execute the clean command with a
relative path.

Also removed the documentation of the moduleRoot option which is really only
for rebasing when running tests.